requirements.txt
singledispatch
numpy
crc
intelhex
xmltodict
requests
pywin32;sys_platform == 'win32'
uptime
pyserial
pyusb
libusb
python-can
pillow
tk
customtkinter
CTkMenuBar
CTkMessagebox
tkinter-async-execute